fix : 끝부분 처리 로직 수정, hiddenSection 제거
